Extract header to its own view
