+++ podcasts/ChaptersViewController.swift
@@ -10,6 +10,12 @@ class ChaptersViewController: PlayerItemViewController {
         }
     }
 
+    private(set) lazy var header: ChaptersHeader = {
+        let header = ChaptersHeader()
+        header.delegate = self
+        return header
+    }()
+
     override func viewDidLoad() {
         super.viewDidLoad()
         chaptersTable.sectionHeaderTopPadding = 0
